requirements.txt
aenum==3.1.11
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.5.1
cryptography==42.0.7
Django==4.2.13
django-cors-headers==4.3.1
django-okta-auth==0.8.0
djangorestframework==3.15.1
drf-yasg==1.21.7
ecdsa==0.19.0
flatdict==4.0.1
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
josepy==1.14.0
mozilla-django-oidc==4.0.1
multidict==6.0.5
okta==2.9.6
okta-jwt==1.3.5
packaging==24.0
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pydash==8.0.1
PyJWT==1.7.1
pyOpenSSL==24.1.0
python-decouple==3.8
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
xmltodict==0.13.0
yarl==1.9.4
python-dotenv>=0.19.0